+++ src/main/java/com/flab/delivery/dto/menugroup/MenuGroupDto.java
@@ -1,4 +1,4 @@
-package com.flab.delivery.dto.menu;
+package com.flab.delivery.dto.menugroup;
 
 
 import com.fasterxml.jackson.annotation.JsonInclude;
