+++ src/main/java/com/flab/delivery/dto/menugroup/MenuGroupRequestDto.java
@@ -1,4 +1,4 @@
-package com.flab.delivery.dto.menu;
+package com.flab.delivery.dto.menugroup;
 
 import lombok.*;
